bug on uploading an existing file solved

+++ src/server/upload.php
@@ -26,10 +26,14 @@ if (isset($_FILES['file'])) {
         echo  json_encode(array(null, array("status" => 400, "message" => $uploaderrors[[$_FILES["file"]["error"]]])));
     } else {
 
-        $destination =  relPath() . $location_file . '/' . basename($_FILES['file']['name']);
-        move_uploaded_file(($_FILES['file']['tmp_name']), $destination);
-        chmod($destination, 0777);
-        $uploadedFile = new File(new SplFileInfo($destination));
-        echo  json_encode(array($uploadedFile, array("status" => 200, "message" => "Success! The file has been uploaded")));
+        $destination =  relPath() . '/' . $location_file . '/' . basename($_FILES['file']['name']);
+        if (file_exists($destination)) echo json_encode(array(null, array("status" => 400, "message" => "Error,  a file named " . $_FILES['file']['name'] . " already exists")));
+
+        else {
+            move_uploaded_file(($_FILES['file']['tmp_name']), $destination);
+            chmod($destination, 0777);
+            $uploadedFile = new File(new SplFileInfo($destination));
+            echo  json_encode(array($uploadedFile, array("status" => 200, "message" => "Success! The file has been uploaded")));
+        }
     }
 }
